fix(routes/cart, models/order) correctly handles having bad animalId being passed

+++ server/db/models/order.js
@@ -28,9 +28,12 @@ Order.getCurrentOrderForUserId = userId => {
 }
 
 Order.prototype.addAnimalQuantity = async function(animalId, quantity) {
+  if (!isNumber(animalId)) {
+    throw Order.AnimalDoesNotExistError
+  }
   const animal = await Animal.findByPk(animalId)
   if (!animal) {
-    throw AnimalDoesNotExistError
+    throw Order.AnimalDoesNotExistError
   }
   const [cartItem, created] = await AnimalOrder.findOrCreate({
     where: {
@@ -54,10 +57,18 @@ Order.prototype.addAnimalQuantity = async function(animalId, quantity) {
   return true
 }
 
+//https://stackoverflow.com/questions/9716468
+function isNumber(n) {
+  return !isNaN(parseFloat(n)) && isFinite(n)
+}
+
 Order.prototype.setAnimalQuantity = async function(animalId, quantity) {
+  if (!isNumber(animalId)) {
+    throw Order.AnimalDoesNotExistError
+  }
   const animal = await Animal.findByPk(animalId)
   if (!animal) {
-    throw AnimalDoesNotExistError
+    throw Order.AnimalDoesNotExistError
   }
   const [cartItem, created] = await AnimalOrder.findOrCreate({
     where: {
@@ -81,9 +92,12 @@ Order.prototype.setAnimalQuantity = async function(animalId, quantity) {
 }
 
 Order.prototype.deleteAnimalOrder = async function(animalId) {
+  if (!isNumber(animalId)) {
+    throw Order.AnimalDoesNotExistError
+  }
   const animal = await Animal.findByPk(animalId)
   if (!animal) {
-    throw AnimalDoesNotExistError
+    throw Order.AnimalDoesNotExistError
   }
   await AnimalOrder.destroy({
     where: {
